Fix CI issues
